refactor(training-metrics): decompose 3,782-line component into modules
Split portfolio/components/ui/Figures/TrainingMetricsAnimation/index.tsx
(3,782 lines) into focused sibling modules, joining the existing
confusionPairs.ts / sparkline.ts pattern. Pure relocation — no logic
rewritten, no rendering branches changed.

  index.tsx                    3,782 -> 181  (orchestrator only)
  synthesisSummaries.ts        1,826  pure synthesis-evidence formatters
  EpochSparkline.tsx             472  sparkline component + layout consts
  SynthesisEvidence.tsx          420  merchant-quality + evidence panels
  ConfusionViz.tsx               338  matrix / top-pairs / heatmap
  TrainingMetricsSkeleton.tsx    302  loading skeleton
  LabelStats.tsx                 210  dataset stats, F1 gauge, label bars
  confusionMatrixView.ts         121  label/matrix view helpers

Verified behavior-preserving: tsc --noEmit clean, 55/55 jest tests pass
(3 suites), eslint clean. The behavior harness (synthesisEvidence.test.tsx)
depends only on the default export rendering identically.
